requirements.txt
streamlit-nightly
pandas
yfinance
akshare
